+++ src/ptree.c
@@ -2,7 +2,6 @@
 #include <Rdefines.h>
 #include <time.h>
 #include "newS4object.h"
-#include "r_memcpy.h"
 
 
 /*
@@ -180,12 +179,12 @@ SEXP R_pncount(SEXP R_x, SEXP R_t, SEXP R_s, SEXP R_o, SEXP R_v) {
     o = INTEGER(PROTECT(allocVector(INTSXP, nr)));
     nprotect++;
     
-    r_memset(o, 0, sizeof(int) * nr);
+    R_chk_memset(o, 0, sizeof(int) * nr);
     
     for (k = 0; k < LENGTH(it); k++)
       o[INTEGER(it)[k]]++;
     
-    r_memset(pb, 0, sizeof(int) * nr);
+    R_chk_memset(pb, 0, sizeof(int) * nr);
     
     for (k = 0; k < LENGTH(ix); k++)
       pb[INTEGER(ix)[k]] = 1;
@@ -242,7 +241,7 @@ SEXP R_pncount(SEXP R_x, SEXP R_t, SEXP R_s, SEXP R_o, SEXP R_v) {
     it = iz;
     
     if (LOGICAL(R_s)[0] == FALSE)
-      r_memcpy(o, pb, sizeof(int) * nr);
+      R_chk_memcpy(o, pb, sizeof(int) * nr);
     
 #ifdef _TIME_H
     t1 = clock();
@@ -285,7 +284,7 @@ SEXP R_pncount(SEXP R_x, SEXP R_t, SEXP R_s, SEXP R_o, SEXP R_v) {
     pnadd(nb[*x], x, n);
     if (LOGICAL(R_s)[0] == FALSE && n > 1) {
       if (n > 2) {
-        r_memcpy(pb, x, sizeof(int) * n);
+        R_chk_memcpy(pb, x, sizeof(int) * n);
         for (k = 0; k < n-1; k++) {
           if (k > 0) {
             j     = pb[0];
@@ -399,7 +398,7 @@ SEXP R_pncount(SEXP R_x, SEXP R_t, SEXP R_s, SEXP R_o, SEXP R_v) {
       INTEGER(r)[i-1] = c;
     else if (n > 1) {
       s = c / t;
-      r_memcpy(pb, x, sizeof(int) * n);
+      R_chk_memcpy(pb, x, sizeof(int) * n);
       for (k = 0; k < n; k++) {
         if (k > 0) {
           j     = pb[0];
@@ -972,7 +971,7 @@ SEXP R_pnrindex(SEXP R_x, SEXP R_v) {
       continue;
     if (n > 1) {
       x = INTEGER(ix)+f;
-      r_memcpy(pb, x, sizeof(int) * n);
+      R_chk_memcpy(pb, x, sizeof(int) * n);
       for (k = 0; k < n; k++) {
         if (k > 0) {
           j     = pb[0];