+++ src/warm.c
@@ -2,7 +2,6 @@
 #include <Rdefines.h>
 #include <time.h>
 #include "newS4object.h"
-#include "r_memcpy.h"
 
 // Iterate a bipartite graph of transactions (itemsets) and
 // items using the HITS algorithm with tolerance FLT_EPSILON.
@@ -69,7 +68,7 @@ SEXP R_hits_ngCMatrix(SEXP x, SEXP R_iter, SEXP R_tol, SEXP R_verbose) {
   z = 0;
   j = INTEGER(R_iter)[0];
   while (j--) {
-    r_memset(z0, 0, sizeof(double) * nr);
+    R_chk_memset(z0, 0, sizeof(double) * nr);
     f = 0;
     for (i = 1; i < nc+1; i++) {
       l = px[i];
@@ -130,7 +129,7 @@ SEXP R_rowWSums_ngCMatrix(SEXP x, SEXP R_weight) {
   ix = getAttrib(x, install("i"));
   
   PROTECT(r = allocVector(REALSXP, nr));
-  r_memset(REAL(r), 0, sizeof(double) * nr);
+  R_chk_memset(REAL(r), 0, sizeof(double) * nr);
   
   f = 0;
   for (i = 1; i < LENGTH(px); i++) {
@@ -457,12 +456,12 @@ SEXP R_weclat_ngCMatrix(SEXP x, SEXP R_weight, SEXP R_support,
   
   SET_VECTOR_ELT(r, 0, (r0 = NEW_OBJECT_OF_CLASS("ngCMatrix")));
   setAttrib(r0, install("p"), PROTECT(r1 = allocVector(INTSXP, n + 1)));
-  r_memcpy(INTEGER(r1), pr, sizeof(int) * (n + 1));
+  R_chk_memcpy(INTEGER(r1), pr, sizeof(int) * (n + 1));
   free(pr); pr = NULL;
   UNPROTECT(1);
   
   setAttrib(r0, install("i"), PROTECT(r1 = allocVector(INTSXP, ni)));
-  r_memcpy(INTEGER(r1), ir, sizeof(int) * ni);
+  R_chk_memcpy(INTEGER(r1), ir, sizeof(int) * ni);
   free(ir); ir = NULL;
   UNPROTECT(1);
   
@@ -471,7 +470,7 @@ SEXP R_weclat_ngCMatrix(SEXP x, SEXP R_weight, SEXP R_support,
   INTEGER(r1)[1] = n;
   
   SET_VECTOR_ELT(r, 1, (r0 = allocVector(REALSXP, n)));
-  r_memcpy(REAL(r0), sr, sizeof(double) * n);
+  R_chk_memcpy(REAL(r0), sr, sizeof(double) * n);
   free(sr); sr = NULL;
   
 #ifdef _TIME_H